+++ internal/webhookca/ca_generator.go
@@ -1,0 +1,67 @@
+package webhookca
+
+import (
+	"bytes"
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
+	"fmt"
+	"math/big"
+	"time"
+
+	core "k8s.io/api/core/v1"
+)
+
+func generateCA(name, namespace string) (map[string][]byte, error) {
+	ca := &x509.Certificate{
+		SerialNumber: big.NewInt(time.Now().Unix()),
+		Subject: pkix.Name{
+			Organization: []string{"Hazelcast"},
+		},
+		NotBefore:             time.Now(),
+		NotAfter:              time.Now().AddDate(10, 0, 0),
+		IsCA:                  true,
+		BasicConstraintsValid: true,
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth},
+		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageCertSign,
+		DNSNames: []string{
+			fmt.Sprintf("%s.%s.svc", name, namespace),
+			fmt.Sprintf("%s.%s.svc.cluster.local", name, namespace),
+		},
+	}
+
+	privateKey, err := rsa.GenerateKey(rand.Reader, 4096)
+	if err != nil {
+		return nil, err
+	}
+
+	rawCrt, err := x509.CreateCertificate(rand.Reader, ca, ca, &privateKey.PublicKey, privateKey)
+	if err != nil {
+		return nil, err
+	}
+
+	var crtPEM bytes.Buffer
+	err = pem.Encode(&crtPEM, &pem.Block{
+		Type:  "CERTIFICATE",
+		Bytes: rawCrt,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	var keyPEM bytes.Buffer
+	err = pem.Encode(&keyPEM, &pem.Block{
+		Type:  "RSA PRIVATE KEY",
+		Bytes: x509.MarshalPKCS1PrivateKey(privateKey),
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return map[string][]byte{
+		core.TLSCertKey:       crtPEM.Bytes(),
+		core.TLSPrivateKeyKey: keyPEM.Bytes(),
+	}, nil
+}
